LXC: Implement RebootInstance of LXCHypervisor
This patch implements RebootInstance function of the LXCHypervisor.
Reboot of the LXC container can be done by using lxc-stop command with
--reboot swtich. Internally this is just sending SIGINT for the init
process of the container.
Since the init process is working on the host machine natively, it
requires CAP_SYS_BOOT capability to call the reboot(2) system call.
This function detects if the CAP_SYS_BOOT is dropped from the container
and raise an error to claim that requirement.
It is up to users if they use the soft reboot of the LXC container with
allowing CAP_SYS_BOOT or drop it and do not use the soft reboot.