LXC: Implement RebootInstance of LXCHypervisor
This patch implements the RebootInstance functionality of the
LXCHypervisor.
Reboot of the LXC container can be done by using the lxc-stop command
with the --reboot switch. Internally this just sends SIGINT to the
init process of the container.
Since the init process is working on the host machine natively, it
requires CAP_SYS_BOOT capability to call the reboot(2) system call.
This function detects if CAP_SYS_BOOT was dropped from the container
and raises an error to enforce that requirement.
It is up to users to decide if they use the soft reboot of the LXC
container by allowing CAP_SYS_BOOT or drop it and do not use the soft
reboot.