Fix CheckPrereq function of the LUInstanceSetParams wrt cpu_mask
Once I try to modify a hvparam of the LXC hypervisor, I got this error:

    ValueError: max() arg is an empty sequence

This is because the current default value of cpu_mask hypervisor
parameter for the LXC hypervisor is the zero-length string, so the
cpu_list which returned by the ParseMultiCpuMask is the blank list.
The zero-length string is a valid for the cpu_mask hvparam, so we need
to skip this check if the cpu_list is the blank list.