LXC: Make the number of ttys a hvparam
This patch adds a new hvparam "lxc_tty" to the LXC hypervisor.
This parameter specifies the number of ttys(actually ptys) that should
be allocated for the container.
The value of lxc.tty should be configurable because in some cases we
want to disable pts instance allocation for the LXC container.